Fixes: Content re-use trial run take 2
Problem: travis build #57 errored -- "Liquid Exception: Included file '_includes/lbaasv1_overview.html' not found in f5-os-lbaasv1/doc/map_install_lbaas_plugin.html
/home/travis/build/jputrino/openstack-f5-lbaasv1/vendor/bundle/ruby/2.1.0/gems/jekyll-2.5.3/lib/jekyll/tags/include.rb:136:in `validate_path': Included file '_includes/lbaasv1_overview.html' not found (IOError)"

Analysis: I moved the images/ dir into lbaasv1/ and copied lbassv1/ into root/_includes

Unit/Robot tests:

Fixes: #5 Missing dependency doc/f5-oslbaasv1-readme.pdf causes build failure.

Problem: Could not build packages because of ref to missing/deprecated pdf file.

Analysis: Update ref to from deprecated pdf file to current html file.

Unit/Robot tests: Successfully built packages.

Fixes: #7 Snat teardown incomplete if two tenants use the same shared subnet.

Problem:

Teardown for a test where a shared network is used by more than one tenant
fails when the shared network/subnet is deleted.

“NetworkInUseClient: Unable to complete operation on network 43f3db3a-40fc-4ef9-9ad2-807115a014f7. There are one or more ports still in use on the network.”

Analysis:

A cache of subnets that are in use by all tenants is being managed by the
agent so that only one set of supporting BIG-IP objects is created for a given
subnet. However, the ref count for the subnet cache is not being properly
managed during snat delete. If more than one tenant uses the same subnet for
LB pool members and all tenants using the subnet tear those objects down
within a short period, the BIG-IP objects associated with the subnet will not
be torn down. Subsequently, during tear down of the shared network/subnet, the
delete fails because there is still a port in use on the subnet for the BIG-IP
objects that were not torn down.

Note: If tenant objects referencing the same subnet are torn down slowly
enough, the periodic cache flush will correct this issue (e.g. set the in use
count to one when there is only one tenant referencing the subnet), and the
tear down will succeed. This issue only occurs when in use count is greater
than one.

Another minor issue encountered while debugging is that the tenant_id
parameter is used as a loop variable.  This can be problematic if this is
not understood and the tenant_id passed in is needed later in the code.  To
avoid confusion for future maintenance, the loop variable was renamed.

Unit/Robot tests:

There is a test in the odk code that triggers this issue by creating/deleting
multiple tenants that have pool member instances on a shared network:

~/odk/python/odk/test/lbaas.py --lightweight-web-server --verbose --optimize --tenant-start 1 --tenant-end 2 --shared-instances 2 --tenant-instances 2