Fixes: I added a branch flag to the git clone command in project_buildScript.sh.
Problem: The _includes and other formatting content was not being copied in as directed by the build script.

Analysis: I think that this may be because that content wasn't in the master branch of the cloned repo, so I added in the -b flag to the clone command and specified 'develop' as the branch to clone.

Unit/Robot tests: